v0: frontend + backend complete! 🎉
Run audio-server and go to http://localhost:1323/v0/ to access the web
UI frontend. This UI provides a "Create Session" button to create music
sessions. When the "create session" button is clicked, a new element is
added to the session list. The session can be given a name, and an ogg
file can be uploaded to the session. Once a file has been given to the
session, the play button next to the session will cause that file to be
uploaded and played.

Many clients can access the server simultaneously. To see additions that
other clients have made, hit the "Refresh List" button.

(See the v0 branch for the complete history)